Delete BasicSkill, Updating PhysicalSkill
